refactor activity timeline to simplify code
