+++ clients/admin-ui/src/features/privacy-requests/events-and-logs/ActivityTimeline.tsx
@@ -4,22 +4,17 @@ import {
   Box,
   useDisclosure,
 } from "fidesui";
+import React, { useEffect, useMemo, useState } from "react";
+
 import {
-  ActivityTimelineItem,
-  ActivityTimelineItemTypeEnum,
   ExecutionLog,
   ExecutionLogStatus,
   PrivacyRequestEntity,
-  PrivacyRequestResults,
-} from "privacy-requests/types";
-import React, { useEffect, useMemo, useState } from "react";
-
-import { formatDate } from "~/features/common/utils";
-import { useGetCommentsQuery } from "~/features/privacy-requests/comments/privacy-request-comments.slice";
-import { CommentResponse } from "~/types/api/models/CommentResponse";
+} from "~/features/privacy-requests/types";
 
 import ActivityTimelineEntry from "./ActivityTimelineEntry";
 import styles from "./ActivityTimelineEntry.module.scss";
+import { usePrivacyRequestComments, usePrivacyRequestEventLogs } from "./hooks";
 import LogDrawer from "./LogDrawer";
 
 type ActivityTimelineProps = {
@@ -38,15 +33,11 @@ const ActivityTimeline = ({ subjectRequest }: ActivityTimelineProps) => {
 
   const { results, id: privacyRequestId } = subjectRequest;
 
-  // Fetch comments data for this privacy request
-  const { data: commentsData, isLoading: isCommentsLoading } =
-    useGetCommentsQuery({
-      privacy_request_id: privacyRequestId,
-      size: 100, // Use a reasonable limit
-    });
-
-  // Determine if results are loading
-  const isResultsLoading = !results;
+  // Use our custom hooks
+  const { commentItems, isLoading: isCommentsLoading } =
+    usePrivacyRequestComments(privacyRequestId);
+  const { eventItems, isLoading: isResultsLoading } =
+    usePrivacyRequestEventLogs(results);
 
   // Combined loading state
   const isLoading = isCommentsLoading || isResultsLoading;
@@ -84,76 +75,30 @@ const ActivityTimeline = ({ subjectRequest }: ActivityTimelineProps) => {
     onOpen();
   };
 
-  // Map from source events to ActivityTimelineItems
-  const mapResultsToTimelineItems = (
-    resultsData?: PrivacyRequestResults,
-  ): ActivityTimelineItem[] => {
-    if (!resultsData) {
-      return [];
-    }
-
-    return Object.entries(resultsData).map(([key, logs]) => {
-      const hasUnresolvedError = logs.some(
-        (log) => log.status === ExecutionLogStatus.ERROR,
-      );
-      const hasSkippedEntry = logs.some(
-        (log) => log.status === ExecutionLogStatus.SKIPPED,
-      );
-
-      return {
-        author: "Fides",
-        title: key,
-        date: formatDate(logs[0].updated_at),
-        type: ActivityTimelineItemTypeEnum.REQUEST_UPDATE,
-        showViewLog: hasUnresolvedError || hasSkippedEntry,
-        onClick: () => showLogs(key, logs),
-        isError: hasUnresolvedError,
-        isSkipped: hasSkippedEntry,
-        id: `request-${key}`,
-      };
-    });
-  };
-
-  // Map comments to ActivityTimelineItem
-  const mapCommentsToTimelineItems = (
-    comments?: CommentResponse[],
-  ): ActivityTimelineItem[] => {
-    if (!comments || comments.length === 0) {
-      return [];
-    }
-
-    return comments.map((comment) => {
-      const author =
-        comment.user_first_name && comment.user_last_name
-          ? `${comment.user_first_name} ${comment.user_last_name}`
-          : comment.username || "Unknown";
-
-      return {
-        author,
-        date: formatDate(comment.created_at),
-        type: ActivityTimelineItemTypeEnum.INTERNAL_COMMENT,
-        showViewLog: false,
-        description: comment.comment_text,
-        isError: false,
-        isSkipped: false,
-        id: `comment-${comment.id}`,
-      };
-    });
-  };
-
   // Combine and sort all timeline items
   const timelineItems = useMemo(() => {
-    const requestItems = mapResultsToTimelineItems(results);
-    const commentItems = mapCommentsToTimelineItems(commentsData?.items);
+    // Override the onClick handler for event items
+    const eventItemsWithClickHandler = eventItems.map((item) => {
+      if (item.type === "Request update" && item.title && results) {
+        const key = item.title;
+        if (results[key]) {
+          return {
+            ...item,
+            onClick: () => showLogs(key, results[key]),
+          };
+        }
+      }
+      return item;
+    });
 
     // Combine both arrays
-    const allItems = [...requestItems, ...commentItems];
+    const allItems = [...eventItemsWithClickHandler, ...commentItems];
 
     // Sort by date (oldest first)
     return allItems.sort((a, b) => {
       return new Date(a.date).getTime() - new Date(b.date).getTime();
     });
-  }, [results, commentsData]);
+  }, [eventItems, commentItems, results]);
 
   // Render skeleton items when loading
   const renderSkeletonItems = () => (
