+++ clients/admin-ui/src/features/privacy-requests/events-and-logs/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from "./usePrivacyRequestComments";
+export * from "./usePrivacyRequestEventLogs";
